changes : revamp all UI of onboarding flow , and api integrated in normal director flow, invite director flow.

+++ src/app/(auth)/login/multi-step-forms/mobileLoginOTPComponents/VerifyMobileOTP.jsx
@@ -1,26 +1,26 @@
-'use client';
-
 import { Button } from '@/components/ui/button';
 import Loading from '@/components/ui/Loading';
+import Slot from '@/components/ui/Slot';
 import { LocalStorageService } from '@/lib/utils';
 import { userVerifyOtp } from '@/services/User_Auth_Service/UserAuthServices';
 import { useMutation } from '@tanstack/react-query';
 import { OTPInput } from 'input-otp';
-import { Clock5 } from 'lucide-react';
+import { ArrowLeft, Clock5 } from 'lucide-react';
+import { useRouter } from 'next/navigation';
+
+import React, { useEffect, useState } from 'react';
 import { toast } from 'sonner';
 import { v4 as uuidv4 } from 'uuid';
-import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
-import Slot from '@/components/ui/Slot';
 
-export default function OTPVerificationForm({ setCurrStep }) {
+const VerifyMobileOTP = ({ setMobileLoginStep }) => {
+  const router = useRouter();
+
   const [startFrom, setStartFrom] = useState(30);
   const [otp, setOtp] = useState();
   const userId = LocalStorageService.get('user_profile');
   const userMobileNumber = LocalStorageService.get('user_mobile_number');
   const operationType = LocalStorageService.get('operation_type');
   const invitationData = LocalStorageService.get('invitationData');
-  const router = useRouter();
 
   useEffect(() => {
     const timer = setInterval(() => {
@@ -46,27 +46,40 @@ export default function OTPVerificationForm({ setCurrStep }) {
         'isEnterpriseOnboardingComplete',
         data.data.data.user.isEnterpriseOnboardingComplete,
       );
+      LocalStorageService.set(
+        'isKycVerified',
+        data.data.data.user.isKycVerified,
+      );
       toast.success('OTP verified successfully');
 
-      // if both userOnboarding & enterpriseOnboarding completed, then redirect to Home Page
+      // if both userOnboarding, enterpriseOnboarding & kyc completed, then redirect to Home Page
       if (
         data.data.data.user.isOnboardingComplete &&
-        data.data.data.user.isEnterpriseOnboardingComplete
+        data.data.data.user.isEnterpriseOnboardingComplete &&
+        data.data.data.user.isKycVerified
       ) {
         const redirectUrl = LocalStorageService.get('redirectUrl');
         LocalStorageService.remove('redirectUrl'); // Clear the redirect URL
         router.push(redirectUrl || '/');
+      }
+      // if userOnboarding & enterpriseOnboarding completed but kyc not verified, then moved to kyc
+      else if (
+        data.data.data.user.isOnboardingComplete &&
+        data.data.data.user.isEnterpriseOnboardingComplete &&
+        !data.data.data.user.isKycVerified
+      ) {
+        router.push('/login/kyc');
       }
       // if userOnboarding is completed but enterpriseOnboarding is not completed, then redirected to enterpriseOnboarding page
       else if (
         data.data.data.user.isOnboardingComplete &&
         !data.data.data.user.isEnterpriseOnboardingComplete
       ) {
-        setCurrStep(4);
+        router.push('/login/enterpriseOnboardingSearch'); // enterprise onboarding search
       }
       // if both userOnboarding and enterpriseOnboarding is inCompleted, then stay in normal flow.
       else {
-        setCurrStep(3);
+        router.push('/login/userOnboarding'); // user onboarding
       }
     },
     onError: (error) => {
@@ -91,19 +104,16 @@ export default function OTPVerificationForm({ setCurrStep }) {
   return (
     <form
       onSubmit={handleVerifiyOTP}
-      className="z-20 flex h-[500px] w-[450px] flex-col items-center justify-center gap-5 rounded-md border border-[#E1E4ED] bg-white p-10"
+      className="flex h-[350px] w-[450px] flex-col items-center justify-center gap-10"
     >
-      <h1 className="w-full text-center text-3xl font-bold text-[#414656]">
-        Welcome to HuesERP!
-      </h1>
-      <p className="w-full text-center text-xl text-[#414656]">
-        One account for all things <span className="font-bold">Hues</span>
-      </p>
-      <h2 className="w-full text-2xl font-bold">Verify OTP</h2>
-      <p className="w-full text-sm">
-        A one time password has been sent to{' '}
-        <span className="font-bold text-[#414656]">+91 {userMobileNumber}</span>
-      </p>
+      <div className="flex flex-col gap-4">
+        <h2 className="w-full text-center text-2xl font-bold">
+          Verify your number
+        </h2>
+        <p className="w-full text-center text-sm font-semibold text-[#A5ABBD]">
+          An one time OTP has been sent to <span>+91 {userMobileNumber}</span>
+        </p>
+      </div>
 
       <OTPInput
         name="otp"
@@ -120,8 +130,8 @@ export default function OTPVerificationForm({ setCurrStep }) {
         )}
       />
 
-      <p className="flex w-full items-center gap-2 text-sm text-[#A5ABBD]">
-        Resend OTP in :{' '}
+      <p className="flex w-full items-center justify-center gap-2 text-sm text-[#A5ABBD]">
+        Resend OTP in{' '}
         <span className="flex items-center gap-1 font-semibold">
           {startFrom >= 0 ? (
             <p className="flex items-center gap-1">
@@ -129,15 +139,27 @@ export default function OTPVerificationForm({ setCurrStep }) {
               00:{startFrom}s
             </p>
           ) : (
-            <Button variant="outline" disabled>
+            <Button size="sm" variant="outline" disabled>
               Resend
             </Button>
           )}
         </span>
       </p>
-      <Button type="Submit" className="w-full">
-        {mutation.isPending ? <Loading /> : 'Submit'}
+      <Button size="sm" type="Submit" className="w-full bg-[#288AF9] p-2">
+        {mutation.isPending ? <Loading /> : 'Verify'}
+      </Button>
+
+      <Button
+        variant="ghost"
+        size="sm"
+        className="w-full p-2"
+        onClick={() => setMobileLoginStep(1)}
+      >
+        <ArrowLeft size={14} />
+        Back
       </Button>
     </form>
   );
-}
+};
+
+export default VerifyMobileOTP;
